+++ tlatools/org.lamport.tlatools/src/tla2sany/semantic/ThmOrAssumpDefNode.java
@@ -460,7 +460,7 @@ public class ThmOrAssumpDefNode extends SymbolNode
       for (int i = 0; i < this.params.length; i++) {
          final Integer plevel = lcSet.get(params[i]);
          if (plevel != null) {
-             this.maxLevels[i] = plevel.intValue();
+             this.maxLevels[i] = plevel;
            }
       }
 
@@ -481,7 +481,7 @@ public class ThmOrAssumpDefNode extends SymbolNode
             this.minMaxLevel[i][j] = MinLevel;
           }
           else {
-            this.minMaxLevel[i][j] = alevel.intValue();
+            this.minMaxLevel[i][j] = alevel;
           }
         }
       }
@@ -600,7 +600,7 @@ public class ThmOrAssumpDefNode extends SymbolNode
 
   @Override
   public final void walkGraph(final Hashtable<Integer, ExploreNode> semNodesTable, final ExplorerVisitor visitor) {
-    final Integer uid = Integer.valueOf(myUID);
+    final Integer uid = myUID;
     if (semNodesTable.get(uid) != null) return;
     semNodesTable.put(uid, this);
     visitor.preVisit(this);
